Separate all jsDependencies-related things in separate projects.
This is part of scala-js/scala-js#2841.

* The contents of `org.scalajs.core.tools.jsdep._` are extracted in
  a separate project `jsdependencies-core`.
* All the settings related to `jsDependencies` are extracted in a
  separate sbt plugin `JSDependenciesPlugin` in
  `jsdependencies-plugin`.
